pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
pydantic = {extras = ["email"], version = "^2.8.0"}
pydantic-settings = "^2"
yarl = "^1.9.2"
ujson = "^5.8.0"
httptools = "^0.6.0"
motor = "^3.5.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.1.3"
colorlog = "^6.8.2"
langchain = "0.1.13"
langchain-community = "0.0.29"
langchain-core = "0.1.33"
langchain-openai = "0.1.1"
langchain-text-splitters = "0.0.1"
langsmith = "0.1.31"
nltk = "3.8.1"
openai = "1.14.2"
poppler-utils = "0.1.0"
pypdf = "4.1.0"
tiktoken = "0.6.0"
pyjwt = "^2.8.0"
pymilvus = "^2.4.4"
python-socketio = "^5.11.3"
requests = "^2.32.3"


[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
httpx = "^0.23.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["server",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "SERVER_ENVIRONMENT=pytest",
    "SERVER_DB_BASE=server_test",
]

[fastapi-template.options]
project_name = "server"
api_type = "rest"
enable_redis = "None"
enable_rmq = "None"
ci_type = "none"
enable_migrations = "None"
enable_taskiq = "None"
enable_kube = "None"
kube_name = "server"
enable_routers = "True"
enable_kafka = "None"
enable_loguru = "None"
traefik_labels = "None"
add_dummy = "None"
orm = "none"
self_hosted_swagger = "None"
prometheus_enabled = "None"
sentry_enabled = "None"
otlp_enabled = "None"
pydanticv1 = "None"
gunicorn = "None"
add_users = "None"
cookie_auth = "None"
jwt_auth = "None"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
